check directory exists or not.

+++ build.cake
@@ -248,7 +248,10 @@ Task("Publish")
 	Environment.SetEnvironmentVariable("TMP",tmp);
 
 	// get the precompiledWeb directory and cake scripts in the desired structure.
-	DeleteDirectory("./artifacts",new  DeleteDirectorySettings(){Recursive=true, Force=true});
+	if (DirectoryExists("./artifacts/"))
+	{
+		DeleteDirectory("./artifacts",new  DeleteDirectorySettings(){Recursive=true, Force=true});
+	}
 	CreateDirectory("./artifacts");
 	CreateDirectory("./artifacts/server");
 	MoveDirectory("./server/precompiledweb","./artifacts/server/precompiledWeb");
